renamed proc_docs to interm_pages
'proc' is not as readable as 'interm' and the purpose of the folder is
to store pages rather than documents.
Also renamed the scripts to say pages and corpus instead of images and
text. remove_anno_paths.py was also renamed to remove_img_path_annos.py.

Fixed up the readme file a bit.